backend/app/database.py
import pymysql
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Database credentials from .env
MYSQL_HOST = os.getenv("MYSQL_HOST", "localhost")
MYSQL_USER = os.getenv("MYSQL_USER", "root")
MYSQL_PASSWORD = os.getenv("MYSQL_PASSWORD", "")
MYSQL_DB = os.getenv("MYSQL_DB", "resturant_chatbot")

def get_db_connection():
    """Creates a database connection and returns the connection object."""
    try:
        conn = pymysql.connect(
            host=MYSQL_HOST,
            user=MYSQL_USER,
            password=MYSQL_PASSWORD,
            database=MYSQL_DB,
            cursorclass=pymysql.cursors.DictCursor
        )
        return conn
    except Exception as e:
        print(f"Failed to connect to database: {e}")
        return None
